Move Navigation Add Page Button from hook to Navigation block
- Moved NavigationAddPageButton component directly into Navigation block edit/index.js
- Added required imports: ToolbarButton, ToolbarGroup, page icon, createBlock, __unstableBlockToolbarLastItem
- Removed old navigation-add-page-button.js hook file
- Removed import from hooks/index.js
- Component now renders directly in Navigation block when isEntityAvailable is true
- Maintains same functionality: only shows in contentOnly mode
